LU-12898 utils: %llu mismatch with type __u64 on ppcle64
Fix build errors like this one on ppcle64:

BUILDSTDERR: libmount_utils_zfs.c: In function 'zfs_mkfs_opts':
BUILDSTDERR: libmount_utils_zfs.c:573:5: error: format '%llu' expects
argument of type 'long long unsigned int', but argument 4 has type
'__u64' [-Werror=format=]
BUILDSTDERR:      mop->mo_device_kb * 1024);

__u64 was treated as an unsigned long long which breaks the build on
ppc64le, where they are not the same size.

In printf cases, cast to unsigned long long to match the printf format
so the format is compatible with the type and it is guaranteed
not to lose any data.

In the case of sscanf(), replace the call with strtoull() to eliminate
the issue.

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/36558
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/utils/libmount_utils_zfs.c
@@ -642,7 +642,7 @@ static char *zfs_mkfs_opts(struct mkfs_opts *mop, char *str, int len)
 		snprintf(str, len, " -o %s", mop->mo_mkfsopts);
 	if (mop->mo_device_kb)
 		snprintf(str, len, " -o quota=%llu",
-			 mop->mo_device_kb * 1024);
+			 (unsigned long long)mop->mo_device_kb * 1024);
 
 	return str;
 }
